+++ src/ruuvi.boards.c/ruuvi_board_gateway.h
@@ -9,7 +9,7 @@
  * Common definitions for Ruuvi Gateway.
  */
 
-#define RB_UART_NRF2ESP         RB_GWBUS_3    //!< UART NRF -> ESP
+#define RB_UART_NRF2ESP         RB_GWBUS_2    //!< UART NRF -> ESP
 #define RB_UART_ESP2NRF         RB_GWBUS_1    //!< UART ESP -> NRF
 #define RB_UART_ESP_RTS         RB_PIN_UNUSED 
 #define RB_UART_ESP_CTS         RB_PIN_UNUSED