+++ src/ruuvi.drivers.c/embedded-sht/embedded-common/sensirion_common.c
@@ -38,7 +38,26 @@
 #include "sensirion_common.h"
 #include "sensirion_i2c.h"
 
-uint8_t sensirion_common_generate_crc(uint8_t *data, uint16_t count) {
+uint16_t sensirion_bytes_to_uint16_t(const uint8_t* bytes) {
+    return (uint16_t)bytes[0] << 8 | (uint16_t)bytes[1];
+}
+
+uint32_t sensirion_bytes_to_uint32_t(const uint8_t* bytes) {
+    return (uint32_t)bytes[0] << 24 | (uint32_t)bytes[1] << 16 |
+           (uint32_t)bytes[2] << 8 | (uint32_t)bytes[3];
+}
+
+float sensirion_bytes_to_float(const uint8_t* bytes) {
+    union {
+        uint32_t u32_value;
+        float float32;
+    } tmp;
+
+    tmp.u32_value = sensirion_bytes_to_uint32_t(bytes);
+    return tmp.float32;
+}
+
+uint8_t sensirion_common_generate_crc(const uint8_t* data, uint16_t count) {
     uint16_t current_byte;
     uint8_t crc = CRC8_INIT;
     uint8_t crc_bit;
@@ -56,15 +75,20 @@ uint8_t sensirion_common_generate_crc(uint8_t *data, uint16_t count) {
     return crc;
 }
 
-int8_t sensirion_common_check_crc(uint8_t *data, uint16_t count,
+int8_t sensirion_common_check_crc(const uint8_t* data, uint16_t count,
                                   uint8_t checksum) {
     if (sensirion_common_generate_crc(data, count) != checksum)
         return STATUS_FAIL;
-    return STATUS_OK;
+    return NO_ERROR;
 }
 
-uint16_t sensirion_fill_cmd_send_buf(uint8_t *buf, uint16_t cmd,
-                                     const uint16_t *args, uint8_t num_args) {
+int16_t sensirion_i2c_general_call_reset(void) {
+    const uint8_t data = 0x06;
+    return sensirion_i2c_write(0, &data, (uint16_t)sizeof(data));
+}
+
+uint16_t sensirion_fill_cmd_send_buf(uint8_t* buf, uint16_t cmd,
+                                     const uint16_t* args, uint8_t num_args) {
     uint8_t crc;
     uint8_t i;
     uint16_t idx = 0;
@@ -76,23 +100,23 @@ uint16_t sensirion_fill_cmd_send_buf(uint8_t *buf, uint16_t cmd,
         buf[idx++] = (uint8_t)((args[i] & 0xFF00) >> 8);
         buf[idx++] = (uint8_t)((args[i] & 0x00FF) >> 0);
 
-        crc = sensirion_common_generate_crc((uint8_t *)&buf[idx - 2],
+        crc = sensirion_common_generate_crc((uint8_t*)&buf[idx - 2],
                                             SENSIRION_WORD_SIZE);
         buf[idx++] = crc;
     }
     return idx;
 }
 
-int16_t sensirion_i2c_read_bytes(uint8_t address, uint8_t *data,
-                                 uint16_t num_words) {
+int16_t sensirion_i2c_read_words_as_bytes(uint8_t address, uint8_t* data,
+                                          uint16_t num_words) {
     int16_t ret;
     uint16_t i, j;
     uint16_t size = num_words * (SENSIRION_WORD_SIZE + CRC8_LEN);
     uint16_t word_buf[SENSIRION_MAX_BUFFER_WORDS];
-    uint8_t *const buf8 = (uint8_t *)word_buf;
+    uint8_t* const buf8 = (uint8_t*)word_buf;
 
     ret = sensirion_i2c_read(address, buf8, size);
-    if (ret != STATUS_OK)
+    if (ret != NO_ERROR)
         return ret;
 
     /* check the CRC for each word */
@@ -100,29 +124,33 @@ int16_t sensirion_i2c_read_bytes(uint8_t address, uint8_t *data,
 
         ret = sensirion_common_check_crc(&buf8[i], SENSIRION_WORD_SIZE,
                                          buf8[i + SENSIRION_WORD_SIZE]);
-        if (ret != STATUS_OK)
+        if (ret != NO_ERROR)
             return ret;
 
         data[j++] = buf8[i];
         data[j++] = buf8[i + 1];
     }
 
-    return STATUS_OK;
+    return NO_ERROR;
 }
 
-int16_t sensirion_i2c_read_words(uint8_t address, uint16_t *data_words,
+int16_t sensirion_i2c_read_words(uint8_t address, uint16_t* data_words,
                                  uint16_t num_words) {
     int16_t ret;
     uint8_t i;
+    const uint8_t* word_bytes;
 
-    ret = sensirion_i2c_read_bytes(address, (uint8_t *)data_words, num_words);
-    if (ret != STATUS_OK)
+    ret = sensirion_i2c_read_words_as_bytes(address, (uint8_t*)data_words,
+                                            num_words);
+    if (ret != NO_ERROR)
         return ret;
 
-    for (i = 0; i < num_words; ++i)
-        data_words[i] = be16_to_cpu(data_words[i]);
+    for (i = 0; i < num_words; ++i) {
+        word_bytes = (uint8_t*)&data_words[i];
+        data_words[i] = ((uint16_t)word_bytes[0] << 8) | word_bytes[1];
+    }
 
-    return STATUS_OK;
+    return NO_ERROR;
 }
 
 int16_t sensirion_i2c_write_cmd(uint8_t address, uint16_t command) {
@@ -133,7 +161,7 @@ int16_t sensirion_i2c_write_cmd(uint8_t address, uint16_t command) {
 }
 
 int16_t sensirion_i2c_write_cmd_with_args(uint8_t address, uint16_t command,
-                                          const uint16_t *data_words,
+                                          const uint16_t* data_words,
                                           uint16_t num_words) {
     uint8_t buf[SENSIRION_MAX_BUFFER_WORDS];
     uint16_t buf_size;
@@ -143,14 +171,14 @@ int16_t sensirion_i2c_write_cmd_with_args(uint8_t address, uint16_t command,
 }
 
 int16_t sensirion_i2c_delayed_read_cmd(uint8_t address, uint16_t cmd,
-                                       uint32_t delay_us, uint16_t *data_words,
+                                       uint32_t delay_us, uint16_t* data_words,
                                        uint16_t num_words) {
     int16_t ret;
     uint8_t buf[SENSIRION_COMMAND_SIZE];
 
     sensirion_fill_cmd_send_buf(buf, cmd, NULL, 0);
     ret = sensirion_i2c_write(address, buf, SENSIRION_COMMAND_SIZE);
-    if (ret != STATUS_OK)
+    if (ret != NO_ERROR)
         return ret;
 
     if (delay_us)
@@ -160,7 +188,7 @@ int16_t sensirion_i2c_delayed_read_cmd(uint8_t address, uint16_t cmd,
 }
 
 int16_t sensirion_i2c_read_cmd(uint8_t address, uint16_t cmd,
-                               uint16_t *data_words, uint16_t num_words) {
+                               uint16_t* data_words, uint16_t num_words) {
     return sensirion_i2c_delayed_read_cmd(address, cmd, 0, data_words,
                                           num_words);
 }