+++ src/ruuvi.drivers.c/src/nrf5_sdk15_platform/communication/ruuvi_nrf5_sdk15_communication_ble_gatt.c
@@ -79,23 +79,25 @@
 #include <stdio.h>
 #include <string.h>
 
-#define CENTRAL_LINK_COUNT               0                                          /**< Number of central links used by the application. When changing this number remember to adjust the RAM settings*/
-#define PERIPHERAL_LINK_COUNT            1                                          /**< Number of peripheral links used by the application. When changing this number remember to adjust the RAM settings*/
+#define CENTRAL_LINK_COUNT               0                     /**< Number of central links used by the application. When changing this number remember to adjust the RAM settings*/
+#define PERIPHERAL_LINK_COUNT            1                     /**< Number of peripheral links used by the application. When changing this number remember to adjust the RAM settings*/
 
-#define APP_BLE_OBSERVER_PRIO            3                                          /**< Application's BLE observer priority. You shouldn't need to modify this value. */
+#define APP_BLE_OBSERVER_PRIO            3                     /**< Application's BLE observer priority. You shouldn't need to modify this value. */
 
-#define FIRST_CONN_PARAMS_UPDATE_DELAY  APP_TIMER_TICKS(5000)                       /**< Time from initiating event (connect or start of notification) to first time sd_ble_gap_conn_param_update is called (5 seconds). */
-#define NEXT_CONN_PARAMS_UPDATE_DELAY   APP_TIMER_TICKS(30000)                      /**< Time between each call to sd_ble_gap_conn_param_update after the first call (30 seconds). */
-#define MAX_CONN_PARAMS_UPDATE_COUNT     3                                          /**< Number of attempts before giving up the connection parameter negotiation. */
+#define NEXT_CONN_PARAMS_UPDATE_DELAY_MS      (30000) /**< Time between each call to sd_ble_gap_conn_param_update after the first call (30 seconds). */
+#define FIRST_CONN_PARAMS_UPDATE_DELAY_TICKS  APP_TIMER_TICKS(5000)  /**< Time from initiating event (connect or start of notification) to first time sd_ble_gap_conn_param_update is called (5 seconds). */
+#define NEXT_CONN_PARAMS_UPDATE_DELAY_TICKS   APP_TIMER_TICKS(NEXT_CONN_PARAMS_UPDATE_DELAY_MS) /**< Time between each call to sd_ble_gap_conn_param_update after the first call (30 seconds). */
 
-#define SEC_PARAM_BOND                   1                                          /**< Perform bonding. */
-#define SEC_PARAM_MITM                   0                                          /**< Man In The Middle protection not required. */
-#define SEC_PARAM_LESC                   0                                          /**< LE Secure Connections not enabled. */
-#define SEC_PARAM_KEYPRESS               0                                          /**< Keypress notifications not enabled. */
-#define SEC_PARAM_IO_CAPABILITIES        BLE_GAP_IO_CAPS_NONE                       /**< No I/O capabilities. */
-#define SEC_PARAM_OOB                    0                                          /**< Out Of Band data not available. */
-#define SEC_PARAM_MIN_KEY_SIZE           7                                          /**< Minimum encryption key size. */
-#define SEC_PARAM_MAX_KEY_SIZE           16                                         /**< Maximum encryption key size. */
+#define MAX_CONN_PARAMS_UPDATE_COUNT     3                     /**< Number of attempts before giving up the connection parameter negotiation. */
+
+#define SEC_PARAM_BOND                   0                     /**< Perform bonding. */
+#define SEC_PARAM_MITM                   0                     /**< Man In The Middle protection not required. */
+#define SEC_PARAM_LESC                   0                     /**< LE Secure Connections not enabled. */
+#define SEC_PARAM_KEYPRESS               0                     /**< Keypress notifications not enabled. */
+#define SEC_PARAM_IO_CAPABILITIES        BLE_GAP_IO_CAPS_NONE  /**< No I/O capabilities. */
+#define SEC_PARAM_OOB                    0                     /**< Out Of Band data not available. */
+#define SEC_PARAM_MIN_KEY_SIZE           7                     /**< Minimum encryption key size. */
+#define SEC_PARAM_MAX_KEY_SIZE           16                    /**< Maximum encryption key size. */
 
 #ifndef RUUVI_NRF5_SDK15_COMMUNICATION_BLE4_GATT_LOG_LEVEL
 #define RUUVI_NRF5_SDK15_COMMUNICATION_BLE4_GATT_LOG_LEVEL RI_LOG_LEVEL_DEBUG
@@ -105,15 +107,21 @@
 #define LOGW(msg) ri_log(RI_LOG_LEVEL_WARNING, msg)
 #define LOGHEX(msg, len) ri_log_hex(RUUVI_NRF5_SDK15_COMMUNICATION_BLE4_GATT_LOG_LEVEL, msg, len)
 
-NRF_BLE_GATT_DEF (m_gatt);                               /**< GATT module instance. */
-NRF_BLE_QWR_DEF (
-    m_qwr);                                  /**< Context for the Queued Write module.*/
-BLE_NUS_DEF (m_nus, NRF_SDH_BLE_TOTAL_LINK_COUNT);       /**< BLE NUS service instance. */
+APP_TIMER_DEF (
+    m_conn_param_retry_timer); //<! Timer for retrying comm param renegotiation.
+
+NRF_BLE_GATT_DEF (m_gatt); /**< GATT module instance. */
+NRF_BLE_QWR_DEF (m_qwr);   /**< Context for the Queued Write module.*/
+BLE_NUS_DEF (m_nus, NRF_SDH_BLE_TOTAL_LINK_COUNT); /**< BLE NUS service instance. */
 static uint16_t m_conn_handle =
     BLE_CONN_HANDLE_INVALID; /**< Handle of the current connection. */
 static bool     m_gatt_is_init = false;
 /**< Pointer to application communication interface, given at initialization */
 static ri_comm_channel_t * channel = NULL;
+
+static ble_gap_conn_params_t m_gatt_params; //!< Holder for GATT params.
+
+static uint8_t m_gatt_retries;
 
 /** @brief Supported PHYs, start at 1MBPS */
 static ble_gap_phys_t m_phys =
@@ -121,14 +129,6 @@ static ble_gap_phys_t m_phys =
     .rx_phys = BLE_GAP_PHY_1MBPS, //BLE_GAP_PHY_2MBPS, BLE_GAP_PHY_CODED
     .tx_phys = BLE_GAP_PHY_1MBPS
 };
-
-// Values selected for optimizing throughput.
-#define MIN_CONN_INTERVAL BLE_GAP_CP_MIN_CONN_INTVL_MIN
-#define MAX_CONN_INTERVAL BLE_GAP_CP_MAX_CONN_INTVL_MIN
-// Apple guideline: MAX_CONN_INTERVAL * SLAVE_LATENCY <= 2 s.
-#define SLAVE_LATENCY     (0U)
-// Apple guideline: MAX_CONN_INTERVAL * (SLAVE_LATENCY + 1) * 3 < CONN_SUP_TIMEOUT
-#define CONN_SUP_TIMEOUT  MSEC_TO_UNITS(8000U, UNIT_10_MS)
 
 /** @brief print PHY enum as string */
 static char const * phy_str (ble_gap_phys_t phys)
@@ -172,10 +172,13 @@ static ret_code_t gap_params_init (void)
     ble_gap_conn_sec_mode_t sec_mode;
     BLE_GAP_CONN_SEC_MODE_SET_OPEN (&sec_mode);
     memset (&gap_conn_params, 0, sizeof (gap_conn_params));
-    gap_conn_params.min_conn_interval = MIN_CONN_INTERVAL;
-    gap_conn_params.max_conn_interval = MAX_CONN_INTERVAL;
-    gap_conn_params.slave_latency     = SLAVE_LATENCY;
-    gap_conn_params.conn_sup_timeout  = CONN_SUP_TIMEOUT;
+    gap_conn_params.min_conn_interval = MSEC_TO_UNITS (RI_GATT_MIN_INTERVAL_STANDARD_MS,
+                                        UNIT_1_25_MS);
+    gap_conn_params.max_conn_interval = MSEC_TO_UNITS (RI_GATT_MAX_INTERVAL_STANDARD_MS,
+                                        UNIT_1_25_MS);
+    gap_conn_params.slave_latency     = RI_GATT_SLAVE_LATENCY_STANDARD;
+    gap_conn_params.conn_sup_timeout  = MSEC_TO_UNITS (RI_GATT_CONN_SUP_TIMEOUT_MS,
+                                        UNIT_10_MS);
     err_code = sd_ble_gap_ppcp_set (&gap_conn_params);
     return err_code;
 }
@@ -222,14 +225,13 @@ static ret_code_t conn_params_init (void)
     ble_conn_params_init_t cp_init;
     memset (&cp_init, 0, sizeof (cp_init));
     cp_init.p_conn_params                  = NULL;
-    cp_init.first_conn_params_update_delay = FIRST_CONN_PARAMS_UPDATE_DELAY;
-    cp_init.next_conn_params_update_delay  = NEXT_CONN_PARAMS_UPDATE_DELAY;
+    cp_init.first_conn_params_update_delay = FIRST_CONN_PARAMS_UPDATE_DELAY_TICKS;
+    cp_init.next_conn_params_update_delay  = NEXT_CONN_PARAMS_UPDATE_DELAY_TICKS;
     cp_init.max_conn_params_update_count   = MAX_CONN_PARAMS_UPDATE_COUNT;
     cp_init.start_on_notify_cccd_handle    = BLE_GATT_HANDLE_INVALID;
     cp_init.disconnect_on_fail             = false;
     cp_init.evt_handler                    = on_conn_params_evt;
     cp_init.error_handler                  = conn_params_error_handler;
-    ble_conn_params_stop(); // added to solve issue 230
     err_code = ble_conn_params_init (&cp_init);
     return err_code;
 }
@@ -288,7 +290,7 @@ static void nus_data_handler (ble_nus_evt_t * p_evt)
  */
 static void ble_evt_handler (ble_evt_t const * p_ble_evt, void * p_context)
 {
-    uint32_t err_code;
+    uint32_t err_code = NRF_SUCCESS;
 
     switch (p_ble_evt->header.evt_id)
     {
@@ -296,6 +298,8 @@ static void ble_evt_handler (ble_evt_t const * p_ble_evt, void * p_context)
             m_conn_handle = p_ble_evt->evt.gap_evt.conn_handle;
             err_code = nrf_ble_qwr_conn_handle_assign (&m_qwr, m_conn_handle);
             LOG ("BLE Connected \r\n");
+            char msg[128];
+            sprintf (msg, "PHY: %s.\r\n", phy_str (m_phys));
             RD_ERROR_CHECK (ruuvi_nrf5_sdk15_to_ruuvi_error (err_code),
                             RD_SUCCESS);
 #           if 0
@@ -319,6 +323,9 @@ static void ble_evt_handler (ble_evt_t const * p_ble_evt, void * p_context)
             evt.type = BLE_NUS_EVT_COMM_STOPPED;
             nus_data_handler (&evt);
             m_conn_handle = BLE_CONN_HANDLE_INVALID;
+            err_code |= app_timer_stop (m_conn_param_retry_timer);
+            RD_ERROR_CHECK (ruuvi_nrf5_sdk15_to_ruuvi_error (err_code),
+                            RD_SUCCESS);
             break;
 
         case BLE_GAP_EVT_PHY_UPDATE_REQUEST:
@@ -392,7 +399,7 @@ static void ble_evt_handler (ble_evt_t const * p_ble_evt, void * p_context)
 
         default:
             // No implementation needed.
-            //LOGD ("BLE Unknown event\r\n");
+            LOGD ("BLE Unknown event\r\n");
             break;
     }
 }
@@ -584,10 +591,10 @@ static ret_code_t peer_manager_init()
     sec_param.oob            = SEC_PARAM_OOB;
     sec_param.min_key_size   = SEC_PARAM_MIN_KEY_SIZE;
     sec_param.max_key_size   = SEC_PARAM_MAX_KEY_SIZE;
-    sec_param.kdist_own.enc  = 1;
-    sec_param.kdist_own.id   = 1;
-    sec_param.kdist_peer.enc = 1;
-    sec_param.kdist_peer.id  = 1;
+    sec_param.kdist_own.enc  = 0;
+    sec_param.kdist_own.id   = 0;
+    sec_param.kdist_peer.enc = 0;
+    sec_param.kdist_peer.id  = 0;
     err_code = pm_sec_params_set (&sec_param);
     err_code |= pm_register (pm_evt_handler);
     return err_code;
@@ -637,6 +644,42 @@ static rd_status_t setup_phys (void)
     return err_code;
 }
 
+static void gatt_params_request (void * const params)
+{
+    ret_code_t nrf_status = NRF_SUCCESS;
+
+    if (BLE_CONN_HANDLE_INVALID == m_conn_handle)
+    {
+        m_gatt_retries = 0;
+        LOG ("No connection, ignore param change\r\n");
+    }
+    else
+    {
+        nrf_status = ble_conn_params_change_conn_params (m_conn_handle, params);
+        LOG ("Requesting param change\r\n");
+    }
+
+    if (NRF_SUCCESS != nrf_status)
+    {
+        nrf_status = app_timer_start (m_conn_param_retry_timer,
+                                      NEXT_CONN_PARAMS_UPDATE_DELAY_TICKS,
+                                      params);
+        RD_ERROR_CHECK (ruuvi_nrf5_sdk15_to_ruuvi_error (nrf_status),
+                        RD_SUCCESS);
+        m_gatt_retries++;
+        LOG ("Param change failed, retry queued\r\n");
+    }
+
+    if (m_gatt_retries > MAX_CONN_PARAMS_UPDATE_COUNT)
+    {
+        // Something is preventing update, disconnect.
+        nrf_status = sd_ble_gap_disconnect (m_conn_handle, BLE_HCI_CONN_INTERVAL_UNACCEPTABLE);
+        RD_ERROR_CHECK (ruuvi_nrf5_sdk15_to_ruuvi_error (nrf_status),
+                        RD_SUCCESS);
+        LOG ("Param change errored too many times, cut connection\r\n");
+    }
+}
+
 rd_status_t ri_gatt_init (void)
 {
     ret_code_t err_code = NRF_SUCCESS;
@@ -658,12 +701,22 @@ rd_status_t ri_gatt_init (void)
         LOGW ("NRF5 SDK15 BLE4 GATT module requires initialized timers\r\n");
         return RD_ERROR_INVALID_STATE;
     }
+    else
+    {
+        // Running timer cannot be recreated.
+        // Stopped timer can be created again, so stop, ignore error, (re)create
+        (void) app_timer_stop (m_conn_param_retry_timer);
+        err_code |= app_timer_create (&m_conn_param_retry_timer, APP_TIMER_MODE_SINGLE_SHOT,
+                                      &gatt_params_request);
+        RD_ERROR_CHECK (ruuvi_nrf5_sdk15_to_ruuvi_error (err_code), ~RD_ERROR_FATAL);
+    }
 
     // Register a handler for BLE events.
     NRF_SDH_BLE_OBSERVER (m_ble_observer, APP_BLE_OBSERVER_PRIO, ble_evt_handler, NULL);
     err_code |= gap_params_init();
     err_code |= nrf_ble_gatt_init (&m_gatt, gatt_evt_handler);
     err_code |= nrf_ble_gatt_att_mtu_periph_set (&m_gatt, NRF_SDH_BLE_GATT_MAX_MTU_SIZE);
+    RD_ERROR_CHECK (ruuvi_nrf5_sdk15_to_ruuvi_error (err_code), ~RD_ERROR_FATAL);
     // Queued Write Module, peer manager cannot be uninitialized, initialize only once.
 
     if (!qwr_is_init)
@@ -672,17 +725,21 @@ rd_status_t ri_gatt_init (void)
         qwr_init.error_handler = nrf_qwr_error_handler;
         err_code |= nrf_ble_qwr_init (&m_qwr, &qwr_init);
         err_code |= peer_manager_init();
+        RD_ERROR_CHECK (ruuvi_nrf5_sdk15_to_ruuvi_error (err_code), ~RD_ERROR_FATAL);
         qwr_is_init = true;
     }
 
     err_code |= conn_params_init();
+    RD_ERROR_CHECK (ruuvi_nrf5_sdk15_to_ruuvi_error (err_code), ~RD_ERROR_FATAL);
+    err_code |= setup_phys();
+    RD_ERROR_CHECK (ruuvi_nrf5_sdk15_to_ruuvi_error (err_code), ~RD_ERROR_FATAL);
 
-    if (NRF_SUCCESS == err_code)
+    if ( (NRF_SUCCESS == err_code))
     {
         m_gatt_is_init = true;
     }
 
-    err_code |= setup_phys();
+    RD_ERROR_CHECK (ruuvi_nrf5_sdk15_to_ruuvi_error (err_code), ~RD_ERROR_FATAL);
     return ruuvi_nrf5_sdk15_to_ruuvi_error (err_code);
 }
 
@@ -847,12 +904,83 @@ rd_status_t ri_gatt_dis_init (const ri_comm_dis_init_t * const p_dis)
     ble_srv_ascii_to_utf8 (&dis_init.serial_num_str, dis_local.deviceid);
     ble_srv_ascii_to_utf8 (&dis_init.hw_rev_str, dis_local.hw_version);
     ble_srv_ascii_to_utf8 (&dis_init.fw_rev_str, dis_local.fw_version);
-    ble_srv_ascii_to_utf8 (&dis_init.sw_rev_str, BUILDNUMBER);
     // Read security level 1, mode 1. OPEN, i.e. anyone can read without encryption.
     // Write not allowed.
     dis_init.dis_char_rd_sec = SEC_OPEN;
     err_code = ble_dis_init (&dis_init);
     return ruuvi_nrf5_sdk15_to_ruuvi_error (err_code);
+}
+
+rd_status_t ri_gatt_params_request (const ri_gatt_params_t params,
+                                    const uint16_t delay_ms)
+{
+    rd_status_t err_code = RD_SUCCESS;
+    ret_code_t nrf_status = NRF_SUCCESS;
+    ble_gap_conn_params_t gap_conn_params = {0};
+    gap_conn_params.conn_sup_timeout = MSEC_TO_UNITS (RI_GATT_CONN_SUP_TIMEOUT_MS,
+                                       UNIT_10_MS);
+
+    switch (params)
+    {
+        case RI_GATT_TURBO:
+            LOG ("RI_GATT_TURBO\r\n");
+            gap_conn_params.slave_latency = RI_GATT_SLAVE_LATENCY_TURBO;
+            gap_conn_params.min_conn_interval = MSEC_TO_UNITS (RI_GATT_MIN_INTERVAL_TURBO_MS,
+                                                UNIT_1_25_MS);
+            gap_conn_params.max_conn_interval = MSEC_TO_UNITS (RI_GATT_MAX_INTERVAL_TURBO_MS,
+                                                UNIT_1_25_MS);
+            break;
+
+        case RI_GATT_LOW_POWER:
+            LOG ("RI_GATT_LOW_POWER\r\n");
+            gap_conn_params.slave_latency = RI_GATT_SLAVE_LATENCY_LOW_POWER;
+            gap_conn_params.min_conn_interval = MSEC_TO_UNITS (RI_GATT_MIN_INTERVAL_LOW_POWER_MS,
+                                                UNIT_1_25_MS);
+            gap_conn_params.max_conn_interval = MSEC_TO_UNITS (RI_GATT_MAX_INTERVAL_LOW_POWER_MS,
+                                                UNIT_1_25_MS);
+            break;
+
+        case RI_GATT_STANDARD:
+        default:
+            LOG ("RI_GATT_STANDARD\r\n");
+            gap_conn_params.slave_latency = RI_GATT_SLAVE_LATENCY_STANDARD;
+            gap_conn_params.min_conn_interval = MSEC_TO_UNITS (RI_GATT_MIN_INTERVAL_STANDARD_MS,
+                                                UNIT_1_25_MS);
+            gap_conn_params.max_conn_interval = MSEC_TO_UNITS (RI_GATT_MAX_INTERVAL_STANDARD_MS,
+                                                UNIT_1_25_MS);
+            break;
+    }
+
+    err_code |= ruuvi_nrf5_sdk15_to_ruuvi_error (app_timer_stop (m_conn_param_retry_timer));
+    memcpy (&m_gatt_params, &gap_conn_params, sizeof (gap_conn_params));
+
+    if (0 == delay_ms)
+    {
+        nrf_status = ble_conn_params_change_conn_params (m_conn_handle, &gap_conn_params);
+
+        if (NRF_SUCCESS != nrf_status)
+        {
+            err_code |=  ruuvi_nrf5_sdk15_to_ruuvi_error (app_timer_start (m_conn_param_retry_timer,
+                         NEXT_CONN_PARAMS_UPDATE_DELAY_TICKS, &m_gatt_params));
+            m_gatt_retries = 1;
+            RD_ERROR_CHECK (err_code, ~RD_ERROR_FATAL);
+            LOG ("Retrying new params soon\r\n");
+        }
+        else
+        {
+            LOG ("Switched to new params\r\n");
+        }
+    }
+    else
+    {
+        err_code |=  ruuvi_nrf5_sdk15_to_ruuvi_error (app_timer_start (m_conn_param_retry_timer,
+                     APP_TIMER_TICKS (delay_ms), &m_gatt_params));
+        m_gatt_retries = 0;
+        RD_ERROR_CHECK (err_code, ~RD_ERROR_FATAL);
+        LOG ("New params set after delay\r\n");
+    }
+
+    return err_code;
 }
 
 #endif