+++ src/ruuvi.drivers.c/src/nrf5_sdk15_platform/flash/ruuvi_nrf5_sdk15_flash.c
@@ -555,5 +555,4 @@ bool ri_flash_is_busy()
     return m_fds_processing;
 }
 
-
 #endif